glusterfs broker

+++ openpaas-service-java-broker-glusterfs/src/main/java/org/openpaas/servicebroker/glusterfs/service/impl/GlusterfsServiceInstanceBindingService.java
@@ -7,6 +7,7 @@ import java.util.UUID;
 
 import org.openpaas.servicebroker.exception.ServiceBrokerException;
 import org.openpaas.servicebroker.exception.ServiceInstanceBindingExistsException;
+import org.openpaas.servicebroker.glusterfs.model.GlusterfsServiceInstance;
 import org.openpaas.servicebroker.model.CreateServiceInstanceBindingRequest;
 import org.openpaas.servicebroker.model.DeleteServiceInstanceBindingRequest;
 import org.openpaas.servicebroker.model.ServiceInstance;
@@ -27,7 +28,7 @@ import org.springframework.stereotype.Service;
  * @author kkanzo@bluedigm.com
  *
  */
-@PropertySource("classpath:datasource.properties")
+@PropertySource("classpath:application-mvc.properties")
 @Service
 public class GlusterfsServiceInstanceBindingService implements ServiceInstanceBindingService {
 
@@ -56,17 +57,11 @@ public class GlusterfsServiceInstanceBindingService implements ServiceInstanceBi
 		logger.debug("GlusterfsServiceInstanceBindingService CLASS createServiceInstanceBinding");
 		
 		/* 최초 ServiceInstanceBinding 생성 요청시에는 해당 ServiceInstanceBinding가 존재하지 않아 해당 메소드를 주석처리 하였습니다.*/
-		//ServiceInstanceBinding binding = glusterfsAdminService.findBindById(request.getBindingId());
 		ServiceInstanceBinding findBinding = glusterfsAdminService.findBindById(request.getBindingId());
 		
 		// 요청 정보로부터 ServiceInstanceBinding 정보를 생성합니다.
 		ServiceInstanceBinding binding = glusterfsAdminService.createServiceInstanceBindingByRequest(request);
 		
-		/* 요청 정보롭터 ServiceInstanceBinding 정보를 생성 할 경우 다음 처리부분이 불필요하여 주석처리 합니다.
-		if (binding != null) {
-			throw new ServiceInstanceBindingExistsException(binding);
-		}
-		*/
 		if(findBinding != null){
 			if(findBinding.getServiceInstanceId().equals(binding.getServiceInstanceId()) &&
 					findBinding.getId().equals(binding.getId())  &&
@@ -82,48 +77,31 @@ public class GlusterfsServiceInstanceBindingService implements ServiceInstanceBi
 		
 		// 요청 정보로부터 ServiceInstance정보를 조회합니다.
 		ServiceInstance instance = glusterfsAdminService.findById(request.getServiceInstanceId());
+		GlusterfsServiceInstance gf = glusterfsAdminService.tenantInfofindById(instance.getServiceInstanceId());
 					
 		// ServiceInstance정보가 엇을경우 예외처리
 		if(instance == null) throw new ServiceBrokerException("Not Exists ServiceInstance");
 		
-		// Database명을 조회합니다.
-		String database = glusterfsAdminService.getDatabase(instance.getServiceInstanceId());
 		// 사용자 아이디를 생성합니다.
 		String username = glusterfsAdminService.getUsername(request.getBindingId());
+		
 		// 사용자 비밀번호를 생성합니다.
-		//String password = UUID.randomUUID().toString().replace("-", "");
 		String password = glusterfsAdminService.getUsername(request.getServiceInstanceId());
 		
-		/* 새로운 사용자명이 존재하는지 검증합니다.*/
-		if (glusterfsAdminService.isExistsUser(username)) {
-			// ensure the instance is empty
-			//glusterfsAdminService.deleteUser(database, username);
-			
-			// 사용자 삭제시 특정 Databas의 사용자를 삭제하지 않아 아래와 같이 사용자 명으로 삭제 처리합니다.
-			glusterfsAdminService.deleteUser(username);
-		}
-		
-		if(glusterfsAdminService.checkUserConnections(instance.getPlanId(), instance.getServiceInstanceId())){
-			throw new ServiceBrokerException("It may not exceed the specified plan.(Not assign Max User Connection)");
-		}
 		// 새로운 사용자를 생성합니다.
-		glusterfsAdminService.createUser(database, username, password);
-		
+		glusterfsAdminService.createUser(gf.getTenantId(), username, password);
+
 		// 반환될 credentials 정보를 생성합니다.
 		Map<String,Object> credentials = new HashMap<String,Object>();
-		credentials.put("name", database);
-		credentials.put("hostname", env.getRequiredProperty("jdbc.host"));
-		credentials.put("port", env.getRequiredProperty("jdbc.port"));
+		credentials.put("tenantname", gf.getTenantName());
+		credentials.put("provider", "swift-keystone");
 		credentials.put("username", username);
 		credentials.put("password", password);
-		credentials.put("uri", glusterfsAdminService.getConnectionString(database, username, password,env.getRequiredProperty("jdbc.host")));
+		credentials.put("auth_url", env.getRequiredProperty("glusterfs.authurl")+"/v2.0");
 		binding = new ServiceInstanceBinding(request.getBindingId(), instance.getServiceInstanceId(), credentials, null, request.getAppGuid());
 		
 		// Binding 정보를 저장합니다.
 		glusterfsAdminService.saveBind(binding);
-		
-		// ServiceInstance의 Plan에 따라 사용자별 MAX_USER_CONNECTIONS 정보를 조정합니다.
-		glusterfsAdminService.setUserConnections(instance.getPlanId(), instance.getServiceInstanceId());
 		
 		return binding;
 	}
@@ -146,13 +124,12 @@ public class GlusterfsServiceInstanceBindingService implements ServiceInstanceBi
 		if(instance ==  null) return null;
 		
 		// bindingId로 사용자를 삭제합니다.
+		if(!binding.getServiceInstanceId().equals(instance.getServiceInstanceId())) return null;
+		
 		glusterfsAdminService.deleteUser(binding.getServiceInstanceId(), bindingId);
 		
 		// bindingId로 Binding 정보를 삭제합니다.
 		glusterfsAdminService.deleteBind(bindingId);
-		
-		// 사용자 삭제휴 ServiceInstance Plan을 기준으로 사용자별 MAX_USER_CONNECTIONS 정보를 조정합니다.
-		glusterfsAdminService.setUserConnections(instance.getPlanId(), instance.getServiceInstanceId());
 		
 		return binding;
 	}
@@ -171,22 +148,21 @@ public class GlusterfsServiceInstanceBindingService implements ServiceInstanceBi
 	 * @return
 	 */
 	public ServiceInstanceBinding getBindingInfo(CreateServiceInstanceBindingRequest request, ServiceInstanceBinding instance){
-		// Database명을 조회합니다.
-		String database = glusterfsAdminService.getDatabase(instance.getServiceInstanceId());
+		
+		GlusterfsServiceInstance gf = glusterfsAdminService.tenantInfofindById(instance.getServiceInstanceId());
+		
 		// 사용자 아이디를 생성합니다.
 		String username = glusterfsAdminService.getUsername(request.getBindingId());
 		// 사용자 비밀번호를 생성합니다.
-		//String password = UUID.randomUUID().toString().replace("-", "");
 		String password = glusterfsAdminService.getUsername(request.getServiceInstanceId());
 		
 		// 반환될 credentials 정보를 생성합니다.
 		Map<String,Object> credentials = new HashMap<String,Object>();
-		credentials.put("name", database);
-		credentials.put("hostname", env.getRequiredProperty("jdbc.host"));
-		credentials.put("port", env.getRequiredProperty("jdbc.port"));
+		credentials.put("tenantname", gf.getTenantName());
+		credentials.put("provider", "swift-keystone");
 		credentials.put("username", username);
 		credentials.put("password", password);
-		credentials.put("uri", glusterfsAdminService.getConnectionString(database, username, password,env.getRequiredProperty("jdbc.host")));
+		credentials.put("auth_url", env.getRequiredProperty("glusterfs.authurl")+"/v2.0");
 		
 		return new ServiceInstanceBinding(request.getBindingId(), instance.getServiceInstanceId(), credentials, null, request.getAppGuid());
